refactor(drift_sync): extract adapters.py (4/8)
Move _labels_diff_to_findings and _protection_diff_to_findings into
adapters.py. Pure functions, no I/O, Finding-producing converters.

__init__.py now imports the adapters, and keeps the checks' supporting
imports (_compute_labels_diff, compute_protection_diff, protection_api,
labels_api, issues_api, GhNotFoundError) which will be moved with the
checks themselves in commit 5.

Refs #47 (Theme A item 4).
